splash/requirements.txt
pillow==8.2.0
flask==2.0.1
